+++ examples/app_host/finite_service.py
@@ -13,7 +13,7 @@ def a_sync_service(service_lifetime: ServiceLifetimeManager):
     print("Service started")
     service_lifetime.set_started()
     print("Service running")
-    time.sleep(10)
+    time.sleep(5)
     print("Service is done")
     # The host should also stop after this point, as all the services have stopped