+++ examples/scheduler/failing_tasks.py
@@ -3,7 +3,7 @@
 import logging
 from datetime import timedelta
 
-from localpost.scheduler import Scheduler, every, delay
+from localpost.scheduler import Scheduler, every, delay, take_first
 
 logging.basicConfig()
 logging.getLogger("localpost").setLevel(logging.DEBUG)
@@ -11,15 +11,15 @@ logging.getLogger("localpost").setLevel(logging.DEBUG)
 scheduler = Scheduler()
 
 
-@scheduler.task(every(timedelta(seconds=3)) // delay((0, 3)))
+@scheduler.task(every(timedelta(seconds=3)) // delay((0, 3)) // take_first(3))
 async def an_async_task():
-    print(f"{an_async_task.__name__} running")
+    print("an_async_task running")
     raise RuntimeError("This is a test error from _async_")
 
 
-@scheduler.task(every(timedelta(seconds=3)) // delay((0, 3)))
+@scheduler.task(every(timedelta(seconds=5)) // delay((0, 3)))
 def a_sync_task():
-    print(f"{a_sync_task.__name__} running")
+    print("a_sync_task running")
     raise RuntimeError("This is a test error from _sync_")