+++ tests/scheduler/cond.py
@@ -3,8 +3,8 @@ from unittest.mock import patch, AsyncMock, Mock
 
 import anyio
 import pytest
-from anyio import Event
 
+from localpost._utils import _Event
 from localpost.scheduler import every, ScheduledTask
 
 pytestmark = pytest.mark.anyio
@@ -18,7 +18,7 @@ async def test_every_trigger():
     # https://docs.python.org/3/library/unittest.mock.html#where-to-patch
     # https://pytest-mock.readthedocs.io/en/latest/usage.html#where-to-patch
     with patch("anyio.sleep", new=AsyncMock(side_effect=_real_anyio_sleep)) as mock_sleep:
-        shutting_down = Event()
+        shutting_down = _Event()
         scheduled_task = Mock(ScheduledTask, shutting_down=shutting_down)
         trigger_factory = scheduled_task_tpl(scheduled_task)
         async with trigger_factory as trigger_events: