+++ tests/scheduler/cond_cron.py
@@ -3,9 +3,9 @@ from unittest.mock import AsyncMock, patch, Mock
 
 import anyio
 import pytest
-from anyio import Event
 from croniter import croniter
 
+from localpost._utils import _Event
 from localpost.scheduler import ScheduledTask
 from localpost.scheduler.cond.cron import cron
 
@@ -22,7 +22,7 @@ async def test_cron_trigger():
     # https://docs.python.org/3/library/unittest.mock.html#where-to-patch
     # https://pytest-mock.readthedocs.io/en/latest/usage.html#where-to-patch
     with patch("anyio.sleep", new=AsyncMock(side_effect=_real_anyio_sleep)) as mock_sleep:
-        shutting_down = Event()
+        shutting_down = _Event()
         scheduled_task = Mock(ScheduledTask, shutting_down=shutting_down)
         trigger_factory = scheduled_task_tpl(scheduled_task)
         async with trigger_factory as trigger_events: