+++ CHANGELOG.md
@@ -7,11 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
 ### Added
 
 ### Changed
 
 ### Removed
+
+## [0.4.0] - 2025-03-23
+
+### Added
+
+- `Host.use()` to apply middlewares
+- Ability to combine multiple hosts (`+` operator)
+
+### Changed
+
+- `EventView.__bool__()` instead of `EventView.is_set()` 
+
+## [0.3.1] - 2025-03-13
+
+### Added
+
+- More tests
 
 ## [0.3.0] - 2025-03-11